Client: Values are Jsons and don't ignore error_codes
Previously Client::get_content returned a sting, now it's returning a
Json object which throws an exception if it's 'null' and when being
casted to a string.  Another issue was that we were not handling errors
from this function.

Note that currently calling this function always fails with a 'key not
found' error. It is because previously this function also waited for the
client to fetch the database at least once. It no longer does so. So a
further fix is still needed.