Insert value as array with IPFS link and parse it back on db query
Only the IPFS content hash is passed back, so the API is kept the same.  That
is, we only produce one ``ipfs:/ipfs/...`` link in a list and we assume that
there is always at least one such link when querying the db, from whence we
extract the plain content hash.