Revert addition of size member to cache entries
IPFS' Cat operation is constant-time (it is basically a lookup for the first
data blob to be retrieved), so it makes no sense to store the data size to
enable computing a timeout dependent on data size.

For the moment, the only operations fetching content use a hardwired timeout
of 1 minute.