Fix build error when cmake re-downloads go-ipfs
The `go get` command doesn't support specifying git commit or tags. So
we must `git checkout` a release version manually. But when cmake
decides to re-download go-ipfs it will try to `git pull` and complaint
that it's not on a branch.

The solution is to do `git checkout master` before doing `go get`.